libmrc: crds - API CHANGE: in crds{_gen} xl an xh are now double
This is a small change that has a big ripple effect. Pretty much everything
in this library has been upgraded to work, but I expect this will break
in external codes.

+++ include/mrc_crds_gen_private.h
@@ -15,8 +15,8 @@ struct mrc_crds_gen {
   
   int n;  // n interior points
   int sw;  // stencil width
-  float xl;  // low edge
-  float xh;  // high edge
+  double xl;  // low edge
+  double xh;  // high edge
 };
 
 struct mrc_crds_gen_ops {
